[Server] Card functions moved to a new module. Added comments and minor changes to index.
